+++ R/deprecate.R
@@ -1,11 +1,8 @@
-
-
 #' @rdname deprecate
 #' @aliases refine-method
 setGeneric(
   name = "refine",
-  def = function(object, ...) standardGeneric("refine"),
-  valueClass = "RefineCA"
+  def = function(object, ...) standardGeneric("refine")
 )
 
 #' @export
@@ -15,8 +12,8 @@ setMethod(
   f = "refine",
   signature = signature(object = "AveragePermutationOrder"),
   definition = function(object, cutoff, margin = c(1, 2), axes = c(1, 2), n = 30) {
-    .Deprecated("bootstrap()", old = "refine()")
-    bootstrap(object, cutoff, margin = margin, axes = axes, n = n)
+    .Deprecated("seriate_refine()", old = "refine()")
+    seriate_refine(object, cutoff, margin = margin, axes = axes, n = n)
   }
 )
 
@@ -27,7 +24,7 @@ setMethod(
   f = "refine",
   signature = signature(object = "BootstrapCA"),
   definition = function(object, cutoff, margin = 1, axes = c(1, 2)) {
-    .Deprecated("bootstrap()", old = "refine()")
-    bootstrap(object, cutoff, margin = margin, axes = axes)
+    .Deprecated("seriate_refine()", old = "refine()")
+    seriate_refine(object, cutoff, margin = margin, axes = axes)
   }
 )