+++ R/reexport.R
@@ -1,1 +1,17 @@
-# REEXPORT
+# REEXPORT FROM ARKHE
+
+#' @importFrom arkhe replace_NA
+#' @export
+arkhe::replace_NA
+
+#' @importFrom arkhe replace_zero
+#' @export
+arkhe::replace_zero
+
+#' @importFrom arkhe remove_NA
+#' @export
+arkhe::remove_NA
+
+#' @importFrom arkhe remove_zero
+#' @export
+arkhe::remove_zero